pyproject.toml
[project]
name = "domashkapro-backend"
version = "0.1.0"
description = ""
authors = [
    { name = "Your Name", email = "you@example.com" },
]
readme = "README.md"
requires-python = ">=3.13,<4.0"

dependencies = [
    "fastapi (>=0.115.12,<0.116.0)",
    "uvicorn[standard] (>=0.34.2,<0.35.0)",
    "pydantic (>=2.11.4,<3.0.0)",
    "httpx (>=0.28.1,<0.29.0)",
    "openai (>=1.78.1,<2.0.0)",
    "python-multipart (>=0.0.20,<0.0.21)",
    "pillow (>=11.2.1,<12.0.0)",
    "python-jose[cryptography] (>=3.4.0,<4.0.0)",
    "passlib[bcrypt] (>=1.7.4,<2.0.0)",
    "email-validator (>=2.2.0,<3.0.0)",
    "slowapi (>=0.1.9,<0.2.0)",
    "sqlalchemy (>=2.0.41,<3.0.0)",
    "aiosqlite (>=0.21.0,<0.22.0)",
    "alembic (>=1.15.2,<2.0.0)",
    "pydantic-settings (>=2.9.1,<3.0.0)",
]

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.5"
pytest-asyncio = "^0.26.0"

[tool.poetry]
package-mode = false

[tool.pytest.ini_options]
asyncio_default_fixture_loop_scope = "function"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
